% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectbraindb.R
\name{insectbraindb_species_info}
\alias{insectbraindb_species_info}
\alias{insectbraindb_neuron_info}
\title{Get information on species/neurons from insectbraindb.org}
\usage{
insectbraindb_species_info()

insectbraindb_neuron_info()
}
\value{
a \code{data.frame} that most importantly gives the user the latin scientific name, common name or neuron names and insectbraindb.org-specific ID number
for each insect species / neuron on which data is held on the site
}
\description{
Get information on the neurons and brain templates hosted by insectbraindb.org, including neuron names, species names, associated publications, etc.
}
\details{
As of May 2019, data from the following \href{https://insectbraindb.org/app/species}{species} is hosted on insectbraindb.org :
\itemize{
 \item 	\emph{Agrotis infusa}	 Bogong moth
 \item 	\emph{Agrotis segetum}	 Turnip moth
 \item 	\emph{Apis mellifera}	 Honeybee
 \item 	\emph{Apis mellifera}	 Honeybee
 \item 	\emph{Danaus plexippus}	 Monarch Butterfly
 \item 	\emph{Helicoverpa armigera}	 Cotton Bollworm,
 \item 	\emph{Helicoverpa assulta}	 Oriental tobacco budworm
 \item 	\emph{Heliothis virescens}	 Tobacco budworm
 \item 	\emph{Macroglossum stellatarum}	 Hummingbird hawk moth
 \item 	\emph{Manduca sexta}	 Tobacco hornworm
 \item 	\emph{Megalopta genalis}	 Sweat bee
 \item 	\emph{Nasonia vitripennis}	 Jewel wasp
 \item 	\emph{Scarabaeus lamarcki}	 Diurnal dung beetle
 \item 	\emph{Schistocerca gregaria}	 Desert Locust
 }
}
\examples{
\dontrun{ 
## What neurons does the insectbraindb.org host?
available.neurons = insectbraindb_neuron_info()

## Let's just download all of the neurons in the database to play with,
## there are not very many:
nrow(available.neurons)

## First, we call the read neurons function, with ids set to NULL
insect.neurons = insectbraindb_read_neurons(ids = NULL)

## Hmm, let's see how many neurons we have perspecies
table(insect.neurons[,"common_name"])

## So, it seem the Monarch Butterfly is the clear winner there, 
## maybe let's just have those
butterfly.neurons = 
subset(insect.neurons, common_name == "Monarch Butterfly")

## And let's plot them
plot3d(butterfly.neurons, lwd = 2, soma = 5)

## Cool! But maybe we also want to see it's template brain? 
## Let's check if they have it
available.brains = insectbraindb_species_info()
available.brains

## Great, they do, let's get it
butterfly.brain = insectbraindb_read_brain(species = "Danaus plexippus")

## And plot in a translucent manner
plot3d(butterfly.brain, alpha = 0.1)

## Oop, that's a lot of neuropils. 
## Let's go for only a subset. What's available?
butterfly.brain$RegionList
butterfly.brain$full_names

# There lateral horn (LH) and the antennal lobe (AL) are my favourites.
# Let's plot those
clear3d()
plot3d(subset(butterfly.brain, "LH|AL"), alpha = 0.5)

}
}
\seealso{
\code{\link{insectbraindb_read_brain}}, \code{\link{insectbraindb_read_neurons}}
}
