% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insectbraindb.R
\name{insectbraindb_read_neurons}
\alias{insectbraindb_read_neurons}
\title{Read insect neurons from insectbraindb.org}
\usage{
insectbraindb_read_neurons(ids = NULL, progress = TRUE)
}
\arguments{
\item{ids}{a neuron ID, or vector of neuron IDs, as recorded in the insectbraindb.org. If set to \code{NULL} (default)
then all neurons in the databse are read. As of May 2018, this is only 100 readable neurons. A helpful information
page on each neuron can be seen by visiting "https://insectbraindb.org/app/neuron/id", where id is a number, e.g.
for Bogon moth neuron \code{219}, this is yuor \href{https://insectbraindb.org/app/neuron/219}{ticket}.}

\item{progress}{if \code{TRUE} or a numeric value, a progress bar is shown to track the state of your download}
}
\value{
a \code{nat} package \code{\link[nat]{neuronlist}} object, replete with metadata, is returned
}
\description{
Read templatebrain-registered neurons from insectbraindb.org (2018 version), given a single ID or vector of neuron IDs. Alternatively,
if \code{ids} is set to \code{NULL}, all neurons in the database that can be read, are read (~100). Metadata for available neurons can be seen, and IDs chosen,
by calling \code{\link{insectbraindb_neuron_info}}. Neurons are returned as a \code{nat} package \code{\link[nat]{neuronlist}} object, 
that can be be plotted in 3D using \code{rgl} and analysed with tools from the \code{nat} ecosystem. Each neuron
in the insectbraindb is represented by a unique numerical ID, a name and is associated with a species of insect and
a publiction, by DOI. Some neurons cannot be read because a SWC file is not available from insectbraindb.org. 
You can examine the available neurons here \href{https://insectbraindb.org/app/neurons}{here}), while the available species
may be examined \href{https://insectbraindb.org/app/species}{here}). Critically, these neurons are registered to standard templates
that may be obtained as \code{hxsurf} objects in R by using \code{\link{insectbraindb_read_brain}}.
}
\details{
Multiple neurons are read, their SWC files as hosted at https://ibdb-file-storage.s3.amazonaws.com/
are downloaded to a temporary directory, and read using \code{\link[nat]{read.neurons}} into a \code{\link[nat]{neuronlist}} object in R.
This format and its manipulation is described in detail \href{https://jefferis.github.io/nat/}{here}.
When using \code{insectbraindb_read_neurons}, meta data for the neuron is also returned that gives a neuron's ID,
short name, long name, associated laboratory and publication, if available, and the assoeciated insect species. As
of May 2019, data from the following \href{https://insectbraindb.org/app/species}{species} is hosted on insectbraindb.org :
\itemize{
 \item 	\emph{Agrotis infusa}	 Bogong moth
 \item 	\emph{Agrotis segetum}	 Turnip moth
 \item 	\emph{Apis mellifera}	 Honeybee
 \item 	\emph{Apis mellifera}	 Honeybee
 \item 	\emph{Danaus plexippus}	 Monarch Butterfly
 \item 	\emph{Helicoverpa armigera}	 Cotton Bollworm,
 \item 	\emph{Helicoverpa assulta}	 Oriental tobacco budworm
 \item 	\emph{Heliothis virescens}	 Tobacco budworm
 \item 	\emph{Macroglossum stellatarum}	 Hummingbird hawk moth
 \item 	\emph{Manduca sexta}	 Tobacco hornworm
 \item 	\emph{Megalopta genalis}	 Sweat bee
 \item 	\emph{Nasonia vitripennis}	 Jewel wasp
 \item 	\emph{Scarabaeus lamarcki}	 Diurnal dung beetle
 \item 	\emph{Schistocerca gregaria}	 Desert Locust
}
 Note that since neurons are reconstructed from many different neural species, 
 there is no 'standard' orientation between species, but within a species these neurons are registered to a
 template brain, usually using elastix. To visualise the templatebrain, use \code{plot3d} on class \code{hxsurf}
 objects downloaded using \code{\link{insectbraindb_read_brain}}.
}
\examples{
\dontrun{ 
## What neurons does the insectbraindb.org host?
available.neurons = insectbraindb_neuron_info()

## Let's just download all of the neurons in the database to play with,
## there are not very many:
nrow(available.neurons)

## First, we call the read neurons function, with ids set to NULL
insect.neurons = insectbraindb_read_neurons(ids = NULL)

## Hmm, let's see how many neurons we have perspecies
table(insect.neurons[,"common_name"])

## So, it seem the Monarch Butterfly is the clear winner there, 
## maybe let's just have those
butterfly.neurons = 
subset(insect.neurons, common_name == "Monarch Butterfly")

## And let's plot them
plot3d(butterfly.neurons, lwd = 2, soma = 5)

## Cool! But maybe we also want to see it's template brain? 
## Let's check if they have it
available.brains = insectbraindb_species_info()
available.brains

## Great, they do, let's get it
butterfly.brain = insectbraindb_read_brain(species = "Danaus plexippus")

## And plot in a translucent manner
plot3d(butterfly.brain, alpha = 0.1)

## Oop, that's a lot of neuropils. 
## Let's go for only a subset. What's available?
butterfly.brain$RegionList
butterfly.brain$full_names

# There lateral horn (LH) and the antennal lobe (AL) are my favourites.
# Let's plot those
clear3d()
plot3d(subset(butterfly.brain, "LH|AL"), alpha = 0.5)

}
}
\seealso{
\code{\link{insectbraindb_neuron_info}}, \code{\link{insectbraindb_read_brain}}, \code{\link{insectbraindb_species_info}}
}
