% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{neuromorpho_async_req}
\alias{neuromorpho_async_req}
\title{Make multiple asynchronous http requests to neuromorpho.org}
\usage{
neuromorpho_async_req(urls, FUN = neuromorphr:::neuromorpho_parse_json,
  batch.size = 10, progress = NULL, message = NULL, ...)
}
\arguments{
\item{urls}{URLs to be queried}

\item{FUN}{a function to run on the responses to requests made using \code{urls}. 
If NULL, the response is returned without any processing. The default is to parse returned JSON format.}

\item{batch.size}{the number of requests sent at once to the neuromorpho.org, using \code{\link[curl]{multi_run}}. 
Requests are sent to neuromorpho.org in parallel to speed up the process of reading neurons. Batches of queries are processed serially.
Increasing the value of \code{batch.size} may reduce read time.}

\item{progress}{if \code{TRUE} or a numeric value, a progress bar is shown. 
The bar progresses when each batch is completed.
If \code{TRUE}, or \code{100}, the bar completes where all batches are done.}

\item{message}{a message to be given alongside progress bar}

\item{...}{methods passed to \code{FUN}}
}
\value{
a list of data retrieved from the requests defined by \code{urls}, and processed by \code{FUN}
}
\description{
Make batches of AJAX style concurrent requests. Used to run multiple asynchronous tasks in parallel
and wait for them all to complete before further batch-wise processing.
}
\details{
Speed querying neuromorpho.org by making batches of AJAX style concurrent requests, 
using using \code{\link[curl]{multi_run}}, and running a custom function on the responses to these requests.
}
\seealso{
\code{\link{neuromorpho_read_neurons}}
}
