% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{neuromorpho_field_counts}
\alias{neuromorpho_field_counts}
\title{Get the number of different entries for a neuromorpho neuron field}
\usage{
neuromorpho_field_counts(field = "species",
  neuromorpho_url = "http://neuromorpho.org", ...)
}
\arguments{
\item{field}{a valid neuron field, as returned by \code{neuromorpho_fields}}

\item{neuromorpho_url}{the base URL for querying the neuromorpho databse, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{neuromorpho_async_req}, or in some cases, \code{neuromorphr:::neuromorpho_fetch}}
}
\description{
Returns a vector of counts for every field entry for a given neuromorpho neuron field.
E.g. you can find out how many neurons there are in the neuromorpho.org repository for each species it has data on.
}
\details{
All the data fields, and their entries, can be seen and explored on neuromorpho.org
at \url{http://neuromorpho.org/MetaData.jsp}. Coutns can be seen as pie charts under the 'browse'
tab at \url{http://neuromorpho.org}, for example, \url{http://neuromorpho.org/byspecies.jsp}.
}
\examples{
\dontrun{
# Get counts for the number of neurons the repository has, by species
species.count = neuromorpho_field_counts(field = "species")
species.count.sorted = sort(species.count,decreasing=TRUE)
species.count.top = c(species.count.top[1:7], 
                      sum(species.count.sorted[8:length(species.count.sorted)]))
names(species.count.top)[length(species.count.top)] = "other"

# Plot
pie(x=species.count.top, labels = names(species.count.top), 
    main="neuron morphologies from different species")

}
}
\seealso{
\code{\link{neuromorpho_search}}, 
\code{\link{neuromorpho_field_entries}},
\code{\link{neuromorpho_fields}}
}
