% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{neuromorpho_field_entries}
\alias{neuromorpho_field_entries}
\title{Return the available meta data entires for neuromorpho neuron fields}
\usage{
neuromorpho_field_entries(field = "species",
  neuromorpho_url = "http://neuromorpho.org", ...)
}
\arguments{
\item{field}{a valid neuron field, as returned by \code{neuromorpho_fields}}

\item{neuromorpho_url}{the base URL for querying the neuromorpho databse, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{\link{neuromorpho_fetch}}}
}
\description{
Returns a list of values present in the repository for the neuron field requested.
These values can be used in the search criteria section of the custom queries.
}
\details{
All the data fields, and their entries, can be seen and explored on neuromorpho.org
at \url{http://neuromorpho.org/MetaData.jsp}.
}
\seealso{
\code{\link{neuromorpho_neuron_info}}, 
\code{\link{neuromorpho_read_neurons}},
\code{\link{neuromorpho_fields}}
}
