% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{neuromorpho_fields}
\alias{neuromorpho_fields}
\title{Return the available meta data fields for neuromorpho neurons}
\usage{
neuromorpho_fields(neuromorpho_url = "http://neuromorpho.org", ...)
}
\arguments{
\item{neuromorpho_url}{the base URL for querying the neuromorpho databse, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{\link{neuromorpho_fetch}}}
}
\description{
Returns a vector of neuron fields available from the neuromorpho.org database.
These fields can be used in the field section of the custom queries.
}
\details{
All the data fields can be seen and explored on neuromorpho.org
at \url{http://neuromorpho.org/MetaData.jsp}.
}
\seealso{
\code{\link{neuromorpho_neuron_info}}, 
\code{\link{neuromorpho_read_neurons}}, 
\code{\link{neuromorpho_field_entries}}
}
