% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literature.R
\name{neuromorpho_get_article}
\alias{neuromorpho_get_article}
\title{Get information on an article by querying an article ID}
\usage{
neuromorpho_get_article(article_id, batch.size = 10,
  neuromorpho_url = "http://neuromorpho.org", progress = TRUE, ...)
}
\arguments{
\item{article_id}{a unique neuromorpho article ID, number in the repository start from 1.
article_ids can be obtained using \code{neuromorpho_articles_from_neurons}}

\item{batch.size}{the number of requests sent at once to the neuromorpho.org, using \code{\link[curl]{multi_run}}. 
Requests are sent to neuromorpho.org in parallel to speed up the process of reading neurons. Batches of queries are processed serially.
Increasing the value of \code{batch.size} may reduce read time.}

\item{neuromorpho_url}{the base URL for querying the neuromorpho databse, defaults to \url{http://neuromorpho.org}}

\item{progress}{if \code{TRUE} or a numeric value, a progress bar is shown. 
The bar progresses when each batch is completed.
If \code{TRUE}, or \code{100}, the bar completes where all batches are done.}

\item{...}{methods passed to \code{neuromorpho_async_req}, or in some cases, \code{neuromorphr:::neuromorpho_fetch}}
}
\description{
Returns a data frame describing an article; article_ids can be obtained using \code{neuromorpho_articles_from_neurons}
}
\details{
Articles can also be searched in a web browser at \url{http://neuromorpho.org/LS.jsp}.
}
\seealso{
Other literature functions: \code{\link{neuromorpho_literature_field_counts}},
  \code{\link{neuromorpho_literature_field_entries}},
  \code{\link{neuromorpho_literature_fields}},
  \code{\link{neuromorpho_literature_search}}
}
\concept{literature functions}
