% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{neuromorpho_is_api_healthy}
\alias{neuromorpho_is_api_healthy}
\title{Check whether the neuromorpho.org API is up and running}
\usage{
neuromorpho_is_api_healthy(neuromorpho_url = "http://neuromorpho.org",
  ...)
}
\arguments{
\item{neuromorpho_url}{the base URL for querying the neuromorpho database, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{neuromorpho_async_req}, or in some cases, \code{neuromorphr:::neuromorpho_fetch}}
}
\value{
TRUE if the API can be reached, FALSE with a warning, if not
}
\description{
Test that the base neuromorpho.org API url, \url{http://neuromorpho.org/api}, can be reached. 
The API provides access to the neuron information, 
morphometry data and literature via three main endpoints 
using conventional HTTP requests. These endpoints are used by function in this package
to read neurons from the repository, get meat information associated with those neurons, and 
perform limited searches of the literature describing neuronal morphologies.
}
\details{
Checks to see whether status the neuromorpho.org API is 'UP' at \url{http://neuromorpho.org/api/health}
}
\examples{
\dontrun{
## Check if the neuromorpho API is healthy
if(neuromorpho_is_api_healthy()){
     message("We can talk to neuromorpho.org!")
}else{
     message("Oh dear")
}
}
}
\seealso{
\code{\link{neuromorpho_search}}, 
\code{\link{neuromorpho_read_neurons}},
\code{\link{neuromorpho_neurons_info}}
}
