% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literature.R
\name{neuromorpho_literature_field_entries}
\alias{neuromorpho_literature_field_entries}
\title{Return the available meta data entries for the literature for neuromorpho neuron}
\usage{
neuromorpho_literature_field_entries(field = "species",
  neuromorpho_url = "http://neuromorpho.org", ...)
}
\arguments{
\item{field}{a valid literature field, as returned by \code{neuromorpho_literature_fields}}

\item{neuromorpho_url}{the base URL for querying the neuromorpho databse, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{neuromorpho_async_req}, or in some cases, \code{neuromorphr:::neuromorpho_fetch}}
}
\description{
Returns a list of values present in the repository for the neuron field requested.
These values can be used in the for custom literature search queries.
}
\details{
Articles can also be searched in a web browser at \url{http://neuromorpho.org/LS.jsp}.
}
\examples{
\dontrun{
# Let's see what the possible entries are for literature fields on neuromorpho.org

## let's have a look at what species are in the literature
lit.fields = neuromorpho_literature_field_entries(field="species")
print(lit.fields)

## Note this is a little different from what we get when we pull this information for neurons
neuron.fields = neuromorpho_field_entries(field="species")
print(neuron.fields)
}

}
\seealso{
Other literature functions: \code{\link{neuromorpho_literature_field_counts}},
  \code{\link{neuromorpho_literature_fields}},
  \code{\link{neuromorpho_literature_search}}
}
\concept{literature functions}
