% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literature.R
\name{neuromorpho_literature_fields}
\alias{neuromorpho_literature_fields}
\title{Return the available meta data fields for the literature for neuromorpho neurons}
\usage{
neuromorpho_literature_fields(neuromorpho_url = "http://neuromorpho.org", ...)
}
\arguments{
\item{neuromorpho_url}{the base URL for querying the neuromorpho database, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{neuromorpho_async_req}, or in some cases, \code{neuromorphr:::neuromorpho_fetch}}
}
\description{
Returns a vector of literature fields available from the neuromorpho.org database.
These fields can be used in the field section of the custom queries.
}
\details{
Articles can also be searched in a web browser at \url{http://neuromorpho.org/LS.jsp}
}
\examples{
# Let's see what meta data fields neuromorpho.org has for its literature database
lit.fields = neuromorpho_literature_fields()
print(lit.fields)

}
\seealso{
Other literature functions: 
\code{\link{neuromorpho_get_article}()},
\code{\link{neuromorpho_literature_field_counts}()},
\code{\link{neuromorpho_literature_field_entries}()},
\code{\link{neuromorpho_literature_search}()}
}
\concept{literature functions}
