% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literature.R
\name{neuromorpho_literature_search}
\alias{neuromorpho_literature_search}
\title{Make a custom literature search to the neuromorpho.org repository}
\usage{

  neuromorpho_literature_search(search_terms = c("species:Drosophila melanogaster"),
  neuromorpho_url = "http://neuromorpho.org", ...)
}
\arguments{
\item{search_terms}{a vector of search terms, with each term formatted as \code{"field:field entry"}.
All valid neuron fields can be seen by calling \code{neuromorpho_fields}, and all the entries for certain fields
can be seen using \code{neuromorpho_fields_entries}. 
Only meta data for neurons that have the entries described for all given search terms will be returned.}

\item{neuromorpho_url}{the base URL for querying the neuromorpho databse, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{\link{neuromorpho_fetch}}}
}
\description{
Search for literature in the neuromorpho.org repository. The returned 
data frame can  be used to get article IDs and neuron names, that can be fed to \code{neuromorpho_get_article}, 
in order to get ifnormation on studies that have submitted reconstructions to neuromorpho.org. 
Searches are made by looking at the meta data for items of literature in the repositroy, 
i.e. querying for certain field entries for certain literature fields.
}
\details{
Beware, that these terms are
different from the equivalent fields and field entries for neurons, e.g. the cell type literature field is \code{cellType} 
rather than \code{cell_type}, as it is for neurons. Articles can also be searched in a web browser at \url{http://neuromorpho.org/LS.jsp}.
}
\examples{
\dontrun{
# Let's search for some literature that has contributed reconstructions to neuromorpho.org

## let's have a look at articles that have worked on Drosophila melanogaster
lit.search = neuromorpho_literature_search(field="species:Drosophila melanogaster")

## We have found this many papers
print(nrow(lit.search))

## Hmm, note that this is a little different from what we get when we pull this information for neurons
neuron.fields = neuromorpho_field_entries(field="species")
print(neuron.fields)
}

}
\seealso{
Other literature functions: \code{\link{neuromorpho_literature_field_counts}},
  \code{\link{neuromorpho_literature_field_entries}},
  \code{\link{neuromorpho_literature_fields}}
}
\concept{literature functions}
