% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphometry.R
\name{neuromorpho_morphometry}
\alias{neuromorpho_morphometry}
\title{Return measurements for queried neurons}
\usage{
neuromorpho_morphometry(neuron_name = NULL, neuron_id = NULL,
  batch.size = 10, data_frame = TRUE,
  neuromorpho_url = "http://neuromorpho.org", progress = TRUE, ...)
}
\arguments{
\item{neuron_name}{a neuron name, or vector of neuron names, as recorded in the neuromorpho database. Names and neuron IDs 
can be found by searching the repository, for example via \code{neuromorpho_search}}

\item{neuron_id}{a neuron ID, or vector of neuron IDs, as recorded in the neuromorpho database. If neuron_name is given
this supersedes \code{neuron_id}, which is then treated as if its value were \code{NULL}.}

\item{batch.size}{the number of requests sent at once to the neuromorpho.org, using \code{\link[curl]{multi_run}}. 
Requests are sent to neuromorpho.org in parallel to speed up the process of reading neurons. Batches of queries are processed serially.
Increasing the value of \code{batch.size} may reduce read time.}

\item{data_frame}{if \code{TRUE}, a \code{data.frame} object is returned, rather than a \code{list} object, with the same information. If 
a nested list (dictionary-like) has been returned for a neuron, this option will result in some data loss.}

\item{neuromorpho_url}{the base URL for querying the neuromorpho databse, defaults to \url{http://neuromorpho.org}}

\item{progress}{if \code{TRUE} or a numeric value, a progress bar is shown. 
The bar progresses when each batch is completed.
If \code{TRUE}, or \code{100}, the bar completes where all batches are done.}

\item{...}{methods passed to \code{neuromorpho_async_req}, or in some cases, \code{neuromorphr:::neuromorpho_fetch}}
}
\value{
a list of measurements for a neuron, or a list of lists if neuron_name/neuron_id is a vector with more than one element
}
\description{
For each neuron corresponding to a given neuron name or neuron ID, return a list of measurements
stored on neuromorpho.org. See details for the sort of measurements you can get.
}
\details{
neurmorpho.org calculates some basic measures for neuronal skeleton reconstructions. Distances are given in μm, and angles in °. Typical measures:
#' \itemize{
  \item neuron_name
  \item neuron_id
  \item surface (μm^2) -- Total arborisation surface area
  \item volume (μm^3) -- Total internal volume of the arborisation
  \item n_branch -- Total number of branches
  \item width  (μm) -- Neuronal width (95% of second principal component)
  \item height (μm) -- Neuronal height (95% of first principal component)
  \item depth (μm) -- Neuronal depth (95% of third principal component)
  \item diameter (μm) -- Average branch diameter
  \item eucDistance (μm) -- Maximum Euclidean (straight) distance from soma to tips
  \item pathDistance -- Maximum Path (along the tree) distance from soma to tips
  \item branch_Order -- Maximum Branch order (number of bifurcations from soma to tips)
  \item contraction -- Average Contraction (the ratio between Euclidean and path length calculated on each branch)
  \item fragmentation -- Total number of reconstruction points
  \item partition_asymmetry -- Topological asymmetry (average over all bifurcations of the absolute value of (n1-n2)/(n1+n2-2), where n1 and n2 are the numbers of tips in the two subtrees)
  \item Rall -- Rall's Ratio (average over all bifurcations of the sum of the diameters of the two daughters, elevated to 1.5, divided by the diameter of the parent, elevated to 1.5)
  \item pk_classic
  \item bif_ampl_local
  \item fractal_Dim
  \item soma_Surface (μm^2) -- Soma surface area
  \item n_stems -- Total number of branches (bifurcations plus terminations)
  \item n_bifs -- Total number of bifurcations
  \item bif_ampl_remote
  \item length (μm) -- Total arborisation length
  \item Local Bifurcation angle (average over all bifurcations of the angle between the first two daughter compartments)
  \item Remote Bifurcation angle (average over all bifurcations of the angle between the following bifurcations or tips)
}

 Neurons may also be searched by their morphological measures via the web browser at \url{http://neuromorpho.org/MorphometrySearch.jsp}

 For exact definitions and/or more information about these measurements, please visit the \href{http://cng.gmu.edu:8080/Lm/help/index.htm}{L-Measure} website. 
 Note that each of these parameters is extracted from the whole neuron (axon plus dendrites).
}
\examples{
# Let's get some measurements for the

# let's get neocortical neurons from both the African elephant and the humpback whale
big.neocortex.df = neuromorpho_search(search_terms= c("species:elephant,humpback whale",
                                                      "brain_region:neocortex"))

## Pull measurements, in a data frame
measurements = neuromorpho_morphometry(big.neocortex.df$neuron_name, data_frame = TRUE)

## Assign species column
measurements$species = big.neocortex.df[rownames(measurements),"species"]

## Boxplot
boxplot(as.numeric(length)~species, data=measurements, notch=FALSE, 
     col=(c("deepskyblue1","firebrick1")),
     main="neocortical neuron lengths", xlab="species")

}
\seealso{
\code{\link{neuromorpho_persistence_vectors}},
\code{\link{neuromorpho_read_neurons}}
}
