% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.R
\name{neuromorpho_names_from_ids}
\alias{neuromorpho_names_from_ids}
\alias{neuromorpho_ids_from_names}
\title{Convert neuromorpho neuron IDs into neuron names, and vica versa}
\usage{
neuromorpho_names_from_ids(neuron_id, progress = TRUE,
  neuromorpho_url = "http://neuromorpho.org", ...)

neuromorpho_ids_from_names(neuron_name, progress = TRUE,
  neuromorpho_url = "http://neuromorpho.org", ...)
}
\arguments{
\item{neuron_id}{a neuron ID, or vector of neuron IDs, as recorded in the neuromorpho database. If neuron_name is given
this supersedes \code{neuron_id}, which is then treated as if its value were \code{NULL}.}

\item{neuromorpho_url}{the base URL for querying the neuromorpho databse, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{\link{neuromorpho_fetch}}}
}
\value{
a vector of neuron names / IDs, corresponding to the given neuron IDs / names
}
\description{
Every neuron in the neuromorpho repository has its own unique neuron ID, starting at 1,
and name. This function maps neuron IDs onto neuron names. Both can be used to read neurons from the repository, using
\code{neuromorpho_read_neurons}.
}
\details{
retrieves the meta data associated with a neuron ID / neuron name, in order to find its corresponding name / ID. 
This is useful because the name can be ued to directly 
locate a neuron's SWC file, but the neuron ID cannot.
}
\examples{
\dontrun{
# find the first 10 neurons uploaded to the neuromorpho respository
neuron_names = neuromorpho_names_from_ids(c(1:10))

# and see that we get the same IDs back if we search the other way
neuron_ids = neuromorpho_ids_from_names(neuron_names)
}
}
\seealso{
\code{\link{neuromorpho_search}}, 
\code{\link{neuromorpho_read_neurons}}
}
