% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{neuromorpho_neurons_info}
\alias{neuromorpho_neurons_info}
\title{Get all meta data associated with a neuron}
\usage{
neuromorpho_neurons_info(
  neuron_name = NULL,
  neuron_id = NULL,
  batch.size = 10,
  progress = TRUE,
  neuromorpho_url = "http://neuromorpho.org",
  ...
)
}
\arguments{
\item{neuron_name}{a neuron name, or vector of neuron names, as recorded in the neuromorpho database. Names and neuron IDs 
can be found by searching the repository, for example via \code{neuromorpho_search}}

\item{neuron_id}{a neuron ID, or vector of neuron IDs, as recorded in the neuromorpho database. If neuron_name is given
this supersedes \code{neuron_id}, which is then treated as if its value were \code{NULL}.}

\item{batch.size}{the number of requests sent at once to the neuromorpho.org, using \code{\link[curl]{multi_run}}. 
Requests are sent to neuromorpho.org in parallel to speed up the process of reading neurons. Batches of queries are processed serially.
Increasing the value of \code{batch.size} may reduce read time.}

\item{progress}{if \code{TRUE} or a numeric value, a progress bar is shown. 
The bar progresses when each batch is completed.
If \code{TRUE}, or \code{100}, the bar completes where all batches are done.}

\item{neuromorpho_url}{the base URL for querying the neuromorpho database, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{neuromorpho_async_req}, or in some cases, \code{neuromorphr:::neuromorpho_fetch}}
}
\value{
a list containing all the meta data for a neuron stored on neuromorpho.org
}
\description{
Retrieve the meta data associated with query neurons. Neurons can be queried using either their
neuron ID or their name.
}
\details{
The meta data for reconstructions has been extracted by neuromorpho.org from corresponding publications, 
and generally include:
\itemize{
  \item Web URL of archives (if available) with any additional information about the reconstruction
  \item Subject-related information - species, strain, age, weight, sex
  \item Experiment-related information - protocol, experimental condition, staining method, slicing direction and thickness, objective type and magnification.
  \item Brain region at three levels: main region/structure, sub-region, detailed sub-region (example: Hippocampus, dentate gyrus, granule cell layer)
  \item Neuron type at three levels: main neuron class type, sub-class, detailed sub-class (example: Interneuron, basket cell, nested)
  \item Reconstruction method
  \item Format of original data, though currently this package only retrieves the data as standardised by neuromorpho.org
}
}
\examples{
\dontrun{ 
# These are some names for a few elephant neurons
elephant_neuron_names = c("155-2-1Stel", "155-2-2Stel", "155-2-5Stel")

# We can now get the meta information associated with these reconstructions
elephant_neuron_info = neuromorpho_neurons_info(elephant_neuron_names)

# We can get essentially the same information as a data frame instead
elephant_neuron_meta = neuromorpho_neurons_meta(elephant_neuron_names)
}
}
\seealso{
\code{\link{neuromorpho_search}}, 
\code{\link{neuromorpho_field_entries}},
\code{\link{neuromorpho_fields}}
}
