% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{neuromorpho_neurons_meta}
\alias{neuromorpho_neurons_meta}
\alias{neuromorpho_neuron_meta}
\title{Get some of the meta data associated with a neuron}
\usage{
neuromorpho_neurons_meta(neuron_name = NULL, neuron_id = NULL,
  light = TRUE, progress = TRUE, batch.size = 10,
  neuromorpho_url = "http://neuromorpho.org", ...)

neuromorpho_neuron_meta(neuron_name = NULL, neuron_id = NULL,
  light = TRUE, neuromorpho_url = "http://neuromorpho.org", ...)
}
\arguments{
\item{neuron_name}{a neuron name, or vector of neuron names, as recorded in the neuromorpho database. Names and neuron IDs 
can be found by searching the repository, for example via \code{neuromorpho_search}}

\item{neuron_id}{a neuron ID, or vector of neuron IDs, as recorded in the neuromorpho database. If neuron_name is given
this supersedes \code{neuron_id}, which is then treated as if its value were \code{NULL}.}

\item{light}{if TRUE, the only a subset of the full meta data for each neurons is returned with the resulting \code{\link[nat]{neuronlist}}.}

\item{progress}{if \code{TRUE} or a numeric value, a progress bar is shown. 
The bar progresses when each batch is completed.
If \code{TRUE}, or \code{100}, the bar completes where all batches are done.}

\item{batch.size}{the number of requests sent at once to the neuromorpho.org, using \code{\link[curl]{multi_run}}. 
Requests are sent to neuromorpho.org in parallel to speed up the process of reading neurons. Batches of queries are processed serially.
Increasing the value of \code{batch.size} may reduce read time.}

\item{neuromorpho_url}{the base URL for querying the neuromorpho database, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{neuromorpho_async_req}, or in some cases, \code{neuromorphr:::neuromorpho_fetch}}
}
\value{
a data frame, with one entry per neuron queried
}
\description{
Retrieve some the meta data associated with query neurons, in a compact way. Neurons can be queried using either their
neuron ID or their name.
}
\details{
The meta data for reconstructions has been extracted by neuromorpho.org from corresponding publications. 
Some of these neuron fields will have multiple entries, 
for example there might be multiple designations for a cell type, or brain region, 
or multiple labs count be credited for a neuron's reconstruction. 
This function returns the majority of the information stored on neuromorpho.org for a given set of neurons, 
but collapses some field entries into one field entry, so we can use a simpler data structure (a data.frame) to represent 
this information, rather than a list. This is useful for adding to \code{neuronlist} objects as an attribute, as is standard in 
\code{nat} family packages. If \code{light = TRUE}, then only the following information is returned in the final data frame:
\itemize{
  \item neuron_id
  \item neuron_name
  \item species
  \item brain_region
  \item cell_type
  \item archive
}
Use \code{\link{neuromorpho_neurons_info}} if you need the full complement of information for a neuron.
}
\examples{
\dontrun{
# These are some names for a few elephant neurons
elephant_neuron_names = c("155-2-1Stel", "155-2-2Stel", "155-2-5Stel")

# We can now get the meta information associated with these reconstructions
elephant_neuron_info = neuromorpho_neurons_info(elephant_neuron_names)

# We can get essentially the same information as a data frame instead
elephant_neuron_meta = neuromorpho_neurons_meta(elephant_neuron_names)
}
}
\seealso{
\code{\link{neuromorpho_search}}, 
\code{\link{neuromorpho_neurons_info}},
\code{\link{neuromorpho_fields}}
}
