% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{neuromorpho_search}
\alias{neuromorpho_search}
\title{Make a custom neuron search to the neuromorpho.org repository}
\usage{
neuromorpho_search(
  search_terms = c("archive:Jacobs", "species:elephant", "brain_region:cerebellum"),
  neuromorpho_url = "http://neuromorpho.org",
  ...
)
}
\arguments{
\item{search_terms}{a vector of search terms, with each term formatted as \code{"field:field entry"}.
All valid neuron fields can be seen by calling \code{neuromorpho_fields}, and all the entries for certain fields
can be seen using \code{neuromorpho_fields_entries}.
Only meta data for neurons that have the entries described for all given search terms will be returned.
To find neurons that are in either of two field entries, one can use, for example, \code{"species:rat,mouse"} to return 
all rat or mouse neurons.}

\item{neuromorpho_url}{the base URL for querying the neuromorpho database, defaults to \url{http://neuromorpho.org}}

\item{...}{methods passed to \code{neuromorpho_async_req}, or in some cases, \code{neuromorphr:::neuromorpho_fetch}}
}
\value{
a data data frame, where each entry is a neuron matching the query specifications.
}
\description{
Search for neurons in the neuromorpho.org repository. The returned 
data frame can  be used to get neuron IDs and neuron names, that can be fed to \code{neuromorpho_read_neurons}, 
in order to read neurons from neuromorpho.org. 
Searches are made by looking at the meta data for neurons in the repositroy, i.e. querying for certain entries in certain neuron fields.
}
\details{
All the data fields, and their entries, can be seen and explored on neuromorpho.org
at \url{http://neuromorpho.org/MetaData.jsp}.
}
\seealso{
\code{\link{neuromorpho_field_entries}}, 
\code{\link{neuromorpho_read_neurons}},
\code{\link{neuromorpho_fields}}
}
