% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz2swc.R
\name{xyz2swc_check}
\alias{xyz2swc_check}
\title{xyz2swc_check: Check SWC files for standard format using xyz2swc tool from NeuroMorpho.org}
\usage{
xyz2swc_check(
  source_dir = NULL,
  source_file = NULL,
  neuromorpho_url = "https://neuromorpho.org",
  ...
)
}
\arguments{
\item{source_dir}{A character string specifying the directory where the source SWC files are located.}

\item{source_file}{A character string specifying the full path to a single source SWC file.}

\item{neuromorpho_url}{A character string specifying the URL for the NeuroMorpho.org website (default is "https://neuromorpho.org").}

\item{...}{Additional arguments.}
}
\value{
A data frame containing a description of any errors in the SWC files.
}
\description{
This function sends SWC files to NeuroMorpho.org so that it can check if those SWC files are in the standard format using its xyz2swc tool. The user gets a data frame describing, for each file, any errors it contains.
}
\examples{
\dontrun{
# check a single file
xyz2swc_check(source_file = "path/to/file.swc")

# check all files in a directory
xyz2swc_check(source_dir = "path/to/directory/")
}
}
\seealso{
\code{\link{xyz2swc_convert}} for a function to convert into standard SWC files.
}
