% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz2swc.R
\name{xyz2swc_convert}
\alias{xyz2swc_convert}
\title{Convert neuron morphology files to standard SWC format}
\usage{
xyz2swc_convert(
  source_dir = NULL,
  source_file = NULL,
  target_zip = file.path(getwd(), "xyz2swc.zip"),
  get_logs = TRUE,
  neuromorpho_url = "https://neuromorpho.org",
  folder = paste0(sample(letters, 10, replace = TRUE), collapse = ""),
  ...
)
}
\arguments{
\item{source_dir}{A character string indicating the directory containing the source files.}

\item{source_file}{A character string indicating the full path to the source file.}

\item{target_zip}{A character string indicating the full path and filename for the resulting .zip file.}

\item{get_logs}{A logical indicating whether to create a log file of any errors or corrections made to the files.}

\item{neuromorpho_url}{A character string indicating the URL of the Neuromorpho website.}

\item{folder}{A character string indicating the name of the folder to store the converted files on the server.}

\item{...}{Additional arguments to pass to `neuromorpho_fetch`.}
}
\value{
A data frame of any errors or corrections made to the files.
}
\description{
This function takes neuron morphology files and converts them into standardised SWC files.
It can take 23 different, common file types as input. The user gets a .zip file downloaded and
a data frame of any errors or corrections made to the files.
}
\examples{
\dontrun{
# Convert a single file
xyz2swc_convert(source_file = "path/to/source/file")

# Convert all files in a directory
xyz2swc_convert(source_dir = "path/to/source/directory")

# Specify a custom URL
xyz2swc_convert(source_file = "path/to/source/file", neuromorpho_url = "https://mycustomurl.com")
}
}
\seealso{
\code{\link{xyz2swc_check}} for a function to validate SWC files.
}
